Moved classes to new structure.
