+++ server/src/main/java/uk/ac/cam/cl/echo/extrusionfinder/server/preprocessor/ImageModifier.java
